fix: start manually triggered scheduled tasks
