+++ client/src/components/apps/tabs/CustomTasksSection.test.jsx
@@ -114,6 +114,7 @@ describe('CustomTasksSection trigger outcomes', () => {
     render(<CustomTasksSection appId="app-1" appName="Example App" />);
     await screen.findByText('Example Task');
     fireEvent.click(screen.getByRole('button', { name: 'Edit' }));
+    expect(screen.queryByLabelText('App scope')).not.toBeInTheDocument();
     fireEvent.change(screen.getByLabelText('Provider'), { target: { value: 'codex' } });
     fireEvent.change(screen.getByLabelText('Model'), { target: { value: 'gpt-5' } });
     fireEvent.change(screen.getByLabelText('Thinking effort'), { target: { value: 'high' } });
@@ -125,6 +126,16 @@ describe('CustomTasksSection trigger outcomes', () => {
       model: 'gpt-5',
       effort: 'high'
     })));
+  });
+
+  it('reports a direct manual trigger as started', async () => {
+    api.triggerCosJob.mockResolvedValue({ success: true, status: 'queued', started: true });
+    render(<CustomTasksSection appId="app-1" appName="Example App" />);
+    await screen.findByText('Example Task');
+
+    fireEvent.click(screen.getByRole('button', { name: 'Run now' }));
+
+    await waitFor(() => expect(toast.success).toHaveBeenCalledWith('Started "Example Task" for Example App'));
   });
 
   it('surfaces a skipped trigger without claiming the task ran', async () => {
